Crowd Control spawned grenades now detonate at the normal speed.
For some reason, if the owner of a grenade is in stasis, the grenade
timer counts down more slowly.  The DXRandoCrowdControlPawn class (which
is used to assign viewer names to things for instigation purposes) will
now always stay out of stasis and move to where the player is when used.
If unused for 5 minutes, the pawn will destroy itself.